+++ utils/local_import.sh
@@ -5,8 +5,11 @@ echo
 echo -e "${txt_blue}  Importing into local database(s).${txt_end}"
 echo
 
-mysql -h$DB_HOST -u$DB_USER -p$DB_PASS $DB_DBASE<$local_backup_dir/latest-m2.sql
+n98-magerun db:import --drop-tables $local_backup_dir/latest-m2.sql
 
 if [ $_arg_wordpress == 'on' ]; then
-    mysql -h$DB_WP_HOST -u$DB_WP_USER -p$DB_WP_PASS $DB_WP_DBASE<$local_backup_dir/latest-wp.sql
+    cd wp
+    sed -i '1s/^/SET SQL_MODE='\''ALLOW_INVALID_DATES'\'';\n\n/' $local_backup_dir/latest-wp.sql
+    
+    wp db import $local_backup_dir/latest-wp.sql
 fi